src/brottsbalken/backend/pyproject.toml
[project]
name = "brottsbalken-backend"
version = "0.1.0"
description = "Backend for Brottsbalken chatbot"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "uvicorn>=0.46.0",
    "lancedb>=0.30.2",
    "pandas>=2.3.3",
    "pydantic>=2.13.3",
    "pydantic-ai>=1.86.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "cohere>=5.21.1",
]

[tool.uv]
package = false
